Adding new distance converters, system.js, and group function
