+++ index.js
@@ -34,4 +34,24 @@ exports.fahrenToCelsius = (fahren, fixed = default_fixed) => {
   return +((fahren - 32.00) / 1.800).toFixed(fixed);
 }
 
+/** 
+ * Converts feet to inches
+ * @param {number} feet feet as a number (ex: 1, 2, 3)
+ * @param {number} fixed the number of decimal places for the conversion
+ * @returns {number} a number in inches
+*/
+exports.feetToInches = (feet, fixed = default_fixed) => {
+  return +(feet * 12).toFixed(fixed);
+}
+
+/** 
+ * Converts inches to feet
+ * @param {number} inches inches as a number (ex: 1, 2, 3)
+ * @param {number} fixed the number of decimal places for the conversion
+ * @returns {number} a number in feet
+*/
+exports.inchesToFeet = (inches, fixed = default_fixed) => {
+  return +(inches / 12).toFixed(fixed);
+}
+
 exports.lhc = lhc;
